Make mounting usb drives easier
I was debating not adding this and using something like `nix-shell` to
add this functionality on the fly when I want to, but when attempting
that I was getting an error when trying to mount the drive/device:

Error looking up object for devices

So instead I enabled this and it can just stay present. This isn't a
service that runs 24/7, there's something that activates it.