startx replaces the shell
I had this weird behavior where SHLVL "started" at 2. By started, I mean
that if I'm in xmonad and open a terminal, echo $SHLVL, I'll see 2. I
find that just a bit odd because it's literally the first shell. This is
due to `exec startx` - we're in a login shell, and exec replaces the zsh
shell with startx. So let's force SHLVL back to 0 here so that it
actually makes sense in future shells :)