Major overhaul of moving neovim to nix!
I'm sure I'm going to miss some details here, but I'd like to point out
a few things:

I got tired of mouse being enabled in neovim. I wish I would have
committed thise prior to all these changes, but here we are.

I didn't realize neovim had its own spellchecker, so I got rid of
spelunker in favor of the built in. That's also why I have `set spell`
in myrc.

Apparently, traces.vim behavior is also default behavior in neovim, so I
was able to get rid of that as well!

I still need to figure out what I'm doing with the remaining plugins in
`plugins-list.vim`

I changed a few plugins as well, such as removed indent-guides which
seemingly weren't even working in favor of indent-blankline.

I'd like to point out that all of the rc or plugin changes made in this
commit were due to 1) an annoyance I've had for awhile or 2) to
accommodate moving to nixos/nixpkgs/nixvim.