I wish I knew about bufdelete when I started vimming!
This `:update` change is completely unrelated, I just happened to
notice it when updating my mapping for the new command for bufdelete.
I'm honestly surprised I just now noticed it! The reason for this change
is that `update` only saves the file if changes have been made, whereas
`write` will always save the file regardless, which can be a little
annoying with file watchers.