nix's default zsh config can be tricky
Apparently the default provided by nixpkgs for /etc/zshrc could conflict
with stuff happening on your own ~/.zshrc - and I found this out the
hard way! Oh my zsh adds stuff to $fpath which causes compinit to need
to get regenerated.

nixpkgs actually warns us about this in the module for zsh and gives us
an option to disable all the default stuff they're doing, but as for now
I think I'll settle for simply disabling global compinit as that was the
culprit.

This was a fun dive with @jfly!